requirements.txt
discord.py==2.3.2
twitchio==2.8.2
httpx==0.27.0
redis==5.0.3
langchain==0.1.16
langchain-community==0.0.36
requests
beautifulsoup4
ollama
python-dotenv
